Beginning refactoring. Moved methods to methods.py
