+++ tests/confindr_report.csv
@@ -6,3 +6,8 @@ Test,Fakella,55,True,22.2,1.1,888,NA
 Test,Fakella,55,True,22.2,1.1,888,NA
 Test,Fakella,55,True,22.2,1.1,888,ND
 Test,Fakella,55,True,22.2,1.1,888,ND
+Test,Fakella,55,True,22.2,1.1,888,ND
+Test,Fakella,55,True,22.2,1.1,888,ND
+Test,Fakella,55,True,22.2,1.1,888,ND
+Test,Fakella,55,True,22.2,1.1,888,ND
+Test,Fakella,55,True,22.2,1.1,888,ND
