+++ tests/example_contamination.csv
@@ -1,24 +1,24 @@
-Gene,Position,Bases,Coverage
-BACT000001_30,1538,T:11;C:54,65
-BACT000002_25,380,C:7;A:43,50
-BACT000005_28,299,C:40;T:18,58
-BACT000006_23,170,T:10;A:57,67
-BACT000009_24,167,T:4;C:40,44
-BACT000011_13,326,C:7;T:26,33
-BACT000011_13,380,T:10;C:3,13
-BACT000015_14,238,G:8;A:35,43
-BACT000019_12,200,C:7;T:21,28
-BACT000030_22,83,T:5;A:25,30
-BACT000030_22,239,G:14;A:57,71
-BACT000030_22,272,T:12;C:53,65
-BACT000030_22,275,T:12;A:53,65
-BACT000030_22,437,G:8;A:52,60
-BACT000031_28,380,T:62;C:18,80
-BACT000031_28,500,C:19;T:57,76
-BACT000031_28,551,C:18;A:49,67
-BACT000032_26,389,G:55;A:17,72
-BACT000038_35,95,G:50;A:7,57
-BACT000040_25,218,T:9;C:54,63
-BACT000044_18,116,G:4;A:35,39
-BACT000044_18,266,T:51;A:10,61
-BACT000051_13,299,T:2;C:27,29
+Gene,Position,SNVCoverage,TotalCoverage
+BACT000001_30,1538,11,65
+BACT000002_25,380,7,50
+BACT000005_28,299,18,58
+BACT000006_23,170,10,67
+BACT000009_24,167,4,44
+BACT000011_13,326,7,33
+BACT000011_13,380,3,13
+BACT000015_14,238,8,43
+BACT000019_12,200,7,28
+BACT000030_22,83,5,30
+BACT000030_22,239,14,71
+BACT000030_22,272,12,65
+BACT000030_22,275,12,65
+BACT000030_22,437,8,60
+BACT000031_28,380,18,80
+BACT000031_28,500,19,76
+BACT000031_28,551,18,67
+BACT000032_26,389,17,72
+BACT000038_35,95,7,57
+BACT000040_25,218,9,63
+BACT000044_18,116,4,39
+BACT000044_18,266,10,61
+BACT000051_13,299,2,29
